Moves Executor into invoke_context.rs
