+++ runtime/src/bank.rs
@@ -70,9 +70,10 @@ use rayon::{
 use solana_measure::measure::Measure;
 use solana_metrics::{inc_new_counter_debug, inc_new_counter_info};
 use solana_program_runtime::{
-    instruction_processor::{Executor, Executors},
     instruction_recorder::InstructionRecorder,
-    invoke_context::{ComputeMeter, ProcessInstructionWithContext, ProgramEntry},
+    invoke_context::{
+        ComputeMeter, Executor, Executors, ProcessInstructionWithContext, ProgramEntry,
+    },
     log_collector::LogCollector,
     timings::ExecuteDetailsTimings,
 };
